5.1- Extracting TypingIndicator Component
