+++ Build-AI-Powered-Apps/4-Project-Building-a-Chatbot/app/packages/client/src/App.tsx
@@ -1,4 +1,4 @@
-import ChatBot from "./components/ChatBot";
+import ChatBot from "./components/chat/ChatBot";
 
 function App() {
   return (
