+++ Build-AI-Powered-Apps/4-Project-Building-a-Chatbot/app/packages/client/src/components/chat/ChatBot.tsx
@@ -3,7 +3,8 @@ import ReactMarkdown from "react-markdown";
 import { useEffect, useRef, useState } from "react";
 import { useForm } from "react-hook-form";
 import { FaArrowUp } from "react-icons/fa";
-import { Button } from "./ui/button";
+import { Button } from "../ui/button";
+import TypingIndicator from "./TypingIndicator";
 
 type FormData = {
   prompt: string;
@@ -92,15 +93,7 @@ const ChatBot = () => {
             <ReactMarkdown>{message.content}</ReactMarkdown>
           </div>
         ))}
-        {isBotTyping && (
-          <div className="self-start">
-            <div className="flex gap-1 px-3 py-3 bg-gray-200 rounded-xl">
-              <div className="w-2 h-2 rounded-full bg-gray-800 animate-pulse"></div>
-              <div className="w-2 h-2 rounded-full bg-gray-800 animate-pulse [animation-delay:0.2s]"></div>
-              <div className="w-2 h-2 rounded-full bg-gray-800 animate-pulse [animation-delay:0.4s]"></div>
-            </div>
-          </div>
-        )}
+        {isBotTyping && <TypingIndicator />}
         {error && <p className="text-red-500">{error}</p>}
       </div>
       <form
