+++ Build-AI-Powered-Apps/4-Project-Building-a-Chatbot/app/packages/client/src/components/chat/TypingIndicator.tsx
@@ -1,0 +1,25 @@
+type DotProps = {
+  className?: string;
+};
+
+const Dot = ({ className }: DotProps) => {
+  return (
+    <div
+      className={`w-2 h-2 rounded-full bg-gray-800 animate-pulse ${className ?? ""}`}
+    />
+  );
+};
+
+const TypingIndicator = () => {
+  return (
+    <div className="self-start">
+      <div className="flex gap-1 px-3 py-3 bg-gray-200 rounded-xl">
+        <Dot />
+        <Dot className="[animation-delay:0.2s]" />
+        <Dot className="[animation-delay:0.4s]" />
+      </div>
+    </div>
+  );
+};
+
+export default TypingIndicator;
